Added --dump functionality
Added an argument -d or --dump which when specified dumps the json to the console rather than into a file. When this new argument is specified, --output is not needed, since we do not want to write to a file when using --dump. When --dump is not used, the else statement on line 71 picks up and restores regular --output functionality. This allows stdout to be captured from the console by other programs, rather than dealing with filesystems, the stdout with --dump can be quickly read and interpreted.